refactor hook
