+++ packages/react-kit/src/hooks/index.ts
@@ -7,3 +7,4 @@ export * from "./useEffectDebugger";
 export * from "./useHandleText";
 export * from "./useExchanges";
 export * from "../components/widgets/finance/useOffersBacked";
+export * from "./contracts/useGetTokenUriImage";
